Reorganize core structure
