+++ touche_rad/core/__init__.py
@@ -1,0 +1,12 @@
+from .context import DebateContext
+from .manager import DebateManager
+from .machine import DebateMachine
+from .strategy import create_strategy, BaseStrategy
+
+__all__ = [
+    "create_strategy",
+    "BaseStrategy",
+    "DebateContext",
+    "DebateManager",
+    "DebateMachine",
+]
