--- touche_rad/core/fsm/__init__.py
@@ -1,9 +1,0 @@
-from .context import DebateContext
-from .manager import DebateManager
-from .machine import DebateMachine
-
-__all__ = [
-    "DebateContext",
-    "DebateManager",
-    "DebateMachine",
-]
