+++ touche_rad/core/strategy/drivers/always_attack.py
@@ -1,5 +1,5 @@
 from .base import BaseStrategy, SystemResponseType
-from touche_rad.core.fsm import DebateContext
+from touche_rad.core.context import DebateContext
 
 
 class AlwaysAttackStrategy(BaseStrategy):
